feat: add /start handler

+++ internal/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID         uint
-	ChatID     uint `gorm:"unique;not null;index"`
-	TelegramID uint `gorm:"unique;not null;index"`
+	ID         int
+	ChatID     int64 `gorm:"unique;not null;index"`
+	TelegramID int64 `gorm:"unique;not null;index"`
 }
